Added FrameRQS back in
needed because ListReport now has column sorting. The link is created
within AFW so the rqs was needed. Problem came up that AFW does not know
if the report is within an ajax or not, so the cp.doc.refreshQueryString
might be for the ajax call, not the original frame. Ecommerce already
maintains a frameRQS though ajax, so that worked.